requirements.txt
flask-apispec
flask-cors
Flask-RESTful
Flask-SQLAlchemy
Flask
